(feat) Add live scratch file merging to CommonStore
The original plan was to have the client perform remote writes
back to both the scratch sheet and the pomsheet, relying on a
combination of revision checking and duplicate checking to
prevent errors. Merging both sheets "live" (ie. in place) works
just as well, though, and it lets me move my practical workflow
forward and focus on higher-level UX, so doing it this way for
now.

Whew, it took me *forever* to get this stupid feature in.