(fix) Always call onPomsheetUpdate to remerge scratch
This is mostly a kludgy sanity-check.The only change is that
the focus handler will now automatically run onPomsheetUpdate()
even if the content_hash from the main file hasn't changed, which
will automatically trigger a re-merge of the scratch sheet.

I'm 99% sure that it won't actually make a difference, because
the reason the scratch sheet isn't updating correctly is that
the longpoll is dying and it's never being fetched when the
scratch sheet updates. But I'm adding it just in case it's
doing something I'm not expecting; also, it's more appropriate
behavior until I fix this mess.

Here's what needs to be done:

1. Make the longpoll in DropboxService work more appropriately
for multiple files (it's the worst code in the codebase, which
is why I've put it off)

2. If I don't feel like doing 1, rewrite checkAndUpdatePomsheet()
(which is the focus handler) to check both files, then call the
appropriate callback.

If this *does* work, I'll write a fix even sooner; before merging
the scratch sheet, updates worked a lot more dependably.