(feat) Allow commentStartHour live-merge from scratch sheet
This isn't really a full solution, but the merge patterning was so
convoluted that I decided to finally take advantage of the fact parsley
doesn't care about multiple date entries. In summary:

1. There's now a commentStartHour type in the typeRegistry, with its
own regex, convert, and parse functions. This gets passed down the
pipeline along with dates and tasks.

2. However, buildMergedPomsheet only handles dates and tasks, and
very laboriously at that. If I want to do it right, but semi-simple,
I'd need to make the merge logic completely agnostic about tasks
and other desiderata, and decide whether to push the desiderata to
the top or bottom of the contents for a given date. I didn't bother
here, though.

3. Instead, at the end of buildMergedPomsheet, it'll just concat
all the commentStartHours that it found in the scratch sheet with
the associated dates and append it to the end of the pomsheet string.