(feat) Add workerized mergeScratchAsync, use it from CommonStore
This commit introduces the following changes:

1. A "skipMerge" param was added to mergeScratch.js to facilitate
workerizing the function. If true, it will return the converted
lines without actually merging the sheet.

2. The file parsley.helper.js was added, with a "plug" util that
helps construct a simplified parsley object before and after
arriving in the worker process.

3. mergeScratchAsync was added, wrapping a worker-spawn and
postMessage in a promise. Before resolving the promise, it calls
the buildMergedScratch from the main thread.

4. CommonStore.js will detect whether Worker is supported, falling
back to a synchronous mergeScratch call if it doesn't.

5. The functions needed by the mergeScratchAsync were moved to
parsley.shared.js so they could be reused. (It was bit arbitrary,
so parsley/index.js should probably be broken up more cleanly.)