(refactor) Move relevant CommonStore code to ParsleyStore and NetworkStore
At long last, I've started breaking out CommonStore. The responsibilites are
as follows, though subject to change:

CommonStore: sort of like a RootStore that has exclusive access to other stores.
Currently also delegates functions in the rest of the app to the broken-out
code, as a stopgap.

NetworkStore: Contains the heartbeat code, $syncBus and other socket handling.
Originally handled login, but moved this to ParsleyStore

ParsleyStore: Login, basic utility functions that aren't strictly related to
the "diegesis" stuff from earlier Vitarka.

The next steps are likely:

1. A VitarkaStore or TurnipStore, to handle and revive some of the book and
  diegesis functionality.
2. Getting $syncBus out of the DropboxService and somewhere more appropriate
3. Possibly setting up a doc-standard RootStore so that stores have access
  to eachother, but I'd strongly prefer to keep them decoupled, so avoiding
  it for now. A very simple eventbus would probably be better.